Enable Rtems tests samples
This patch enables rtems tests samples on the do_one script.
When setting up the environment for testing stdlibc++ you need this test
samples to check everything it's ok before running the gcc test suite.
This check consists on running a sample (i.e ticker.exe) making sure it
behaves as it should.